tonomi-scripts/scripted-components/requirements.txt
unittest
requests_mock
pyyaml
mock